core: fixed error in LuaState::getLuaVersion() for version 5.1, which didn't define the constants I was using

+++ .github/workflows/actions/test_core/entrypoint.sh
@@ -3,12 +3,29 @@
 cd src/core
 LUA_VERSION=5.1 cmake -DCMAKE_BUILD_TYPE=Debug .
 
+if [ 0 -ne $? ]; then
+    exit 1
+fi
+
 # Make sure the build for core succeeds
 make -j2 trogdor
 
+if [ 0 -ne $? ]; then
+    exit 1
+fi
+
 # Build and run the unit tests for core against Lua 5.1
 make -j2 test_core
+
+if [ 0 -ne $? ]; then
+    exit 1
+fi
+
 ./test_core
+
+if [ 0 -ne $? ]; then
+    exit 1
+fi
 
 make clean
 rm -rf CMakeFiles && rm CMakeCache.txt
@@ -17,12 +34,29 @@ rm -rf CMakeFiles && rm CMakeCache.txt
 
 LUA_VERSION=5.2 cmake -DCMAKE_BUILD_TYPE=Debug .
 
+if [ 0 -ne $? ]; then
+    exit 1
+fi
+
 # Make sure the build for core succeeds
 make -j2 trogdor
 
+if [ 0 -ne $? ]; then
+    exit 1
+fi
+
 # Build and run the unit tests for core against Lua 5.2
 make -j2 test_core
+
+if [ 0 -ne $? ]; then
+    exit 1
+fi
+
 ./test_core
+
+if [ 0 -ne $? ]; then
+    exit 1
+fi
 
 make clean
 rm -rf CMakeFiles && rm CMakeCache.txt
@@ -31,9 +65,26 @@ rm -rf CMakeFiles && rm CMakeCache.txt
 
 LUA_VERSION=5.3 cmake -DCMAKE_BUILD_TYPE=Debug .
 
+if [ 0 -ne $? ]; then
+    exit 1
+fi
+
 # Make sure the build for core succeeds
 make -j2 trogdor
 
+if [ 0 -ne $? ]; then
+    exit 1
+fi
+
 # Build and run the unit tests for core against Lua 5.3
 make -j2 test_core
+
+if [ 0 -ne $? ]; then
+    exit 1
+fi
+
 ./test_core
+
+if [ 0 -ne $? ]; then
+    exit 1
+fi
